// src/app/shared/Models/UpdateCustomer.ts
export class UpdateCustomer {          
    title?: string | null;               
    firstName!: string | null;           
    middleName?: string | null;         
    lastName!: string |null;           
    suffix!: string |null;             
    companyName?: string |null;                 
    emailAddress?: string |null;      
    phone?: string |null;              
  }
  